WeatherApi/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
python-dotenv==1.0.0
requests==2.31.0
Flask-Limiter==3.3.1
ipinfo==4.4.2
flask-swagger-ui==4.11.1
pyOpenSSL==23.2.0